Pubsub_firestore_trigger/requirements.txt
functions-framework
google-cloud-firestor